% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{distill_pivot_par}
\alias{distill_pivot_par}
\title{Parallelized Stable Distillation for the OLS Model}
\usage{
distill_pivot_par(y, x, Q, max_num_causal, p_order = seq_len(ncol(x)), ...)
}
\arguments{
\item{y}{a n x m matrix of outcome vectors, one outcome per column}

\item{x}{a n x p matrix of predictors (may be passed as a \code{sparseMatrix}, see the \code{Matrix} package)}

\item{Q}{an orthogonal matrix whose columns span the column space of the background covariates}

\item{max_num_causal}{a non-negative integer, maximum number of causal/active predictors to search for}

\item{p_order}{a vector of non-negative integers, if provided, columns of \code{x} will be distilled in the order \code{p_order[1]},\code{p_order[2]},..}
}
\value{
a list containing

  \code{p_value} a vector of m p-values from the Renyi Outlier Test, in the same order as the outcome vectors provided

  \code{u} a p x m matrix of extracted p-values

  \code{signs} a p x m matrix of effect signs corresponding the extracted p-values \code{u}. Note, these signs are useful for post-hoc interpretation but are NOT extracted in the SD procedure so they are NOT independent of \code{u}.

  \code{y} a n x p matrix of distilled outcome vectors y left over as a biproduct of distillation (independent of \code{u}) and ready for downstream testing
}
\description{
Distill multiple outcome vectors under the same null and alternative OLS model (in parallel)
}
\details{
Parallelized implementation of the OLS distillation routine with simple quantile filter proposed in Section 4 of \link{https://arxiv.org/pdf/2212.12539} .
}
