% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{FindTargetVars}
\alias{FindTargetVars}
\title{Helper for Finding Target Variants for Screening}
\usage{
FindTargetVars(
  map,
  min.cM = 0.1,
  from = 1L,
  to = length(map),
  initial.targets = NULL,
  smt.thresh = NULL
)
}
\arguments{
\item{map}{a vector specifying the recombination map 'CDF' at all cached variants}

\item{min.cM}{the desired minimum recombination distance between target loci in cM}

\item{from}{the first variant index to consider as a target}

\item{to}{the last variant index to consider as a target}

\item{initial.targets}{an integer vector or data.frame.
If an integer vector, the indices of variants that should be included in the returned targets.
If a \code{data.frame}, single marker test results in the format returned by \code{\link{TestCachedMarkers}} to be used
in combination with \code{smt.thresh} to implicitly define initial target variants}

\item{smt.thresh}{when \code{initial.targets} is provided as a \code{data.frame} of single marker test results,
a -log10 p-value threshold such that variants with any signal above that threshold should be included in initial targets}
}
\value{
integer vector of target variant indices
}
\description{
Use recombination distance (and preliminary single marker test results) to select target variants for screening
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{TestCachedMarkers}} to select variants based on initial single marker test results
}
