% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestMarker.R
\name{TestCachedMarkers}
\alias{TestCachedMarkers}
\title{Test All Markers
Fast method for performing SMT on a genotype matrix}
\usage{
TestCachedMarkers(
  y,
  A = NULL,
  from = 1L,
  to = kalis::L(),
  ploidy = 2L,
  model = "additive"
)
}
\description{
Test All Markers
Fast method for performing SMT on a genotype matrix
}
