% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestClades.R
\name{TestCladeMat}
\alias{TestCladeMat}
\title{Test a Clade Matrix for association with phenotypes}
\usage{
TestCladeMat(y, M, Q, traces = NULL)
}
\arguments{
\item{y}{a matrix of phenotypes}

\item{M}{a matrix as produced by \code{kalis::CladesMat}}

\item{Q}{an orthogonal matrix whose columns span the column space of the background covariates}

\item{traces}{an optional traces argument}
}
\value{
With with p-values for Clade Testing
}
\description{
Test a local phylogeny for association with phenotypes based on a matrix prodced by \code{kalis::CladesMat}
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{Clades}} to generate \code{kalisClades} object
}
