% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestMarker.R
\name{TestMarker}
\alias{TestMarker}
\title{Perform SMT}
\usage{
TestMarker(x, g, add.noise = FALSE)
}
\arguments{
\item{x}{a null model object as returned by \link{FitNull}}

\item{g}{a genotype vector}

\item{add.noise}{if 'raw' return raw residuals, if FALSE simply rank normalize residuals, if TRUE add structured noise along the subspace spanned by the background covs and genotype so that the residuals are 'full rank': truly from N(0,I), not a curved Gaussian.}
}
\description{
Perform SMT
}
