% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestSprigs.R
\name{TestSprigs}
\alias{TestSprigs}
\title{Test Sprigs}
\usage{
TestSprigs(
  y,
  sprigs,
  ortho = FALSE,
  Q = matrix(1/sqrt(nrow(y)), nrow(y), 1),
  use.forking = FALSE
)
}
\arguments{
\item{y}{a \code{n} x \code{m} matrix of phenotype residuals, one phenotype per column}

\item{sprigs}{a list as returned by \link{\code{kalis::Sprigs}}}

\item{ortho}{a logical, to be deprecated, currently kept for background compatibility}

\item{Q}{a \code{n} x \code{q} orthogonal matrix with columns spanning the column space of the background covariates}

\item{use.forking}{a logical, is multiprocessing by process forking allowed?  Some relatively minor acclerations are possible if TRUE, but users should verify that it is safe to launch forked processes on their compute cluster.}
}
\description{
Test Sprigs object as returned by \link{\code{kalis::Sprigs}} using Stable Distillation.
}
