% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestMarker.R
\name{FitNull}
\alias{FitNull}
\title{Fit Null Linear Model}
\usage{
FitNull(y, A = NULL)
}
\arguments{
\item{y}{a n x m matrix of m phenotypes (one phenotype / column)}

\item{A}{a n x q matrix of q background covariates}
}
\description{
Fit Null Linear Model
}
