% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msse.R
\name{msse.test}
\alias{msse.test}
\title{MSSE Test
Test MSSE}
\usage{
msse.test(x, y, z, test.1.solo = TRUE)
}
\arguments{
\item{x}{vector of p-values}

\item{y}{vector of p-values}

\item{z}{vector of p-values}
}
\description{
MSSE Test
Test MSSE
}
