% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestClades.R
\name{TestCladeMat}
\alias{TestCladeMat}
\title{Test a Clade Matrix for association with phenotypes}
\usage{
TestCladeMat(
  y,
  M,
  Q,
  traces = NULL,
  other.test.pvalues = NULL,
  min.prop.var = 0.98,
  k = c(10, 100, 200, 400, 800),
  neg.log10.cutoff = NULL,
  use.bettermc = FALSE,
  use.forking = FALSE,
  nthreads = 1L
)
}
\arguments{
\item{y}{a matrix of phenotypes}

\item{M}{a matrix as produced by \code{kalis::CladesMat}}

\item{Q}{an orthogonal matrix whose columns span the column space of the background covariates}

\item{traces}{an optional traces argument}

\item{point.est}{a logical, if TRUE evaluate eigenvalues according to schedule k until at least 99% of the variation in M is captured and return the resulting p-value.
If FALSE (default), stop evaluation of eigenvalues once bounds can rule out that locus is significant}
}
\value{
With with p-values for Clade Testing
}
\description{
Test a local phylogeny for association with phenotypes based on a matrix prodced by \code{kalis::CladesMat}
}
\examples{
\dontrun{
}

}
\seealso{
\code{\link{Clades}} to generate \code{kalisClades} object
}
